feat(default ticket blocks): Make all ticket blocks optional
By default, for new _events_ we have been loading the tickets, rsvp, and attendees blocks. This has caused some extra work for folks in the EU. The attendees block in particular is an "always on" block - so once it's added it shows. As part of making things better for GDPR folks, I made all the blocks optional and selectable in the backend.

:movie_camera: http://p.tri.be/1Mbt0N
:movie_camera: http://p.tri.be/OOGYgf

BREAKING CHANGE: Attendee block no longer loads automatically for new events. However this is now controllable from the Admin events->settings->tickets screen.

:ticket: https://central.tri.be/issues/138887